add render comments method into popup component
